Bugfix: Feed watchdog within busy-wait-loop within connectBearSSL to prevent a premature reset. (#157)
Within the Arduino IoT Cloud firmware stack we are using a watchdog to prevent unintended application lockup. Unfortunately in certain situations the connection attempt undertaken within WiFiNINA can take so long that the watchdog is triggered although everything is fine. This change is feeding the watchdog (if one is present) so to avoid prematurely triggering of the watchdog.