Fix address of NTP server in example (#188)
For some reason the example failed to get a response from the IP address of `time.nist.gov`. I also didn't get a ping reply from that address to my laptop.
I had more luck with one of the servers of `pool.ntp.org`.
For the record I'm located in Amsterdam, Netherlands. Does this new address work in US and other parts of the world?